// src/app/shared/services/model/model.ts
export interface testimonials {
  name: string;
  location:string;
  title: string;
  content: string;
  
  
}
export interface thumbnails1 {
  img: string;
}